requirements.txt
djangorestframework==3.15.2
django-configurations==2.5.1
django-cors-headers==4.7.0
django-filter==25.1
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.10
str2bool==1.1
